+++ cli/src/commands/operation/show.rs
@@ -37,6 +37,19 @@ pub struct OperationShowArgs {
     /// Don't show the graph, show a flat list of modified changes
     #[arg(long)]
     no_graph: bool,
+    /// Render the operation using the given template
+    ///
+    /// You can specify arbitrary template expressions using the
+    /// [built-in keywords]. See [`jj help -k templates`] for more
+    /// information.
+    ///
+    /// [built-in keywords]:
+    ///     https://jj-vcs.github.io/jj/latest/templates/#operation-keywords
+    ///
+    /// [`jj help -k templates`]:
+    ///     https://jj-vcs.github.io/jj/latest/templates/
+    #[arg(long, short = 'T', add = ArgValueCandidates::new(complete::template_aliases))]
+    template: Option<String>,
     /// Show patch of modifications to changes
     ///
     /// If the previous version has different parents, it will be temporarily
@@ -44,6 +57,9 @@ pub struct OperationShowArgs {
     /// contaminated by unrelated changes.
     #[arg(long, short = 'p')]
     patch: bool,
+    /// Do not show operation diff
+    #[arg(long, conflicts_with_all = ["patch", "DiffFormatArgs"])]
+    no_op_diff: bool,
     #[command(flatten)]
     diff_format: DiffFormatArgs,
 }
@@ -88,9 +104,11 @@ pub fn cmd_op_show(
         })
     };
 
-    // TODO: Should we make this customizable via clap arg?
     let template: TemplateRenderer<Operation> = {
-        let text = settings.get_string("templates.op_log")?;
+        let text = match &args.template {
+            Some(value) => value.to_owned(),
+            None => settings.get_string("templates.op_log")?,
+        };
         workspace_command
             .parse_operation_template(ui, &text)?
             .labeled(["op_show", "operation"])
@@ -100,20 +118,23 @@ pub fn cmd_op_show(
     let mut formatter = ui.stdout_formatter();
     template.format(&op, formatter.as_mut())?;
 
-    // TODO: Merged repo may have newly rebased commits, which wouldn't exist in
-    // the index. (#4465)
-    if parent_ops.len() > 1 {
-        return Ok(());
+    if !args.no_op_diff {
+        // TODO: Merged repo may have newly rebased commits, which wouldn't exist in
+        // the index. (#4465)
+        if parent_ops.len() > 1 {
+            return Ok(());
+        }
+        show_op_diff(
+            ui,
+            formatter.as_mut(),
+            repo.as_ref(),
+            &parent_repo,
+            &repo,
+            &commit_summary_template,
+            (!args.no_graph).then_some(graph_style),
+            &with_content_format,
+            diff_renderer.as_ref(),
+        )?;
     }
-    show_op_diff(
-        ui,
-        formatter.as_mut(),
-        repo.as_ref(),
-        &parent_repo,
-        &repo,
-        &commit_summary_template,
-        (!args.no_graph).then_some(graph_style),
-        &with_content_format,
-        diff_renderer.as_ref(),
-    )
+    Ok(())
 }